converted hubot to use fifo queues
This strategy doesn't actually work as well as desired.
It does improve performance but ends up creating a big
kinda-distributed pipeline.  Any failure of a plugin
causes everything to crash, and without notice.

The issue is that a plugin exiting causes the fifo to
break (broken pipe error), so that tee cannot write to
all of it's targets.